+++ CHANGELOG.md
@@ -777,3 +777,4 @@
 - [2026-04-29 19:21] feat(storage): integrate custom telemetry logger in storage
 - [2026-04-29 19:37] feat(forms): integrate custom telemetry logger in forms
 - [2026-04-29 20:24] feat(core): add async core stream provider
+- [2026-04-29 20:25] refactor(cache): migrate cache controller to modern notifier syntax